extract vue compile icon function
