+++ components/icon/vue/index.ts
@@ -1,6 +1,7 @@
 import type { iconsMetadata } from '@cypress-design/icon-registry'
 export type { WindiColor } from '@cypress-design/icon-registry'
-export { default, compileVueIconProperties } from './Icon'
+export { default } from './Icon'
+export { compileVueIconProperties } from './compileProperties'
 export * from './TreeShakableIcons'
 
 export type IconId = keyof typeof iconsMetadata
